🔧 refactor(student): reorganize class info fetch API under 'student' module
